feat: Creating a custom error class

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,6 @@
 import AsyncStorage from "@react-native-async-storage/async-storage"; 
+import { AppError } from "@utils/AppError";
+
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 
 import { groupsGetAll } from "./groupsGetAll"
@@ -7,6 +9,12 @@ export async function groupCreate(newGroup: string){
   try{
 
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists){
+      throw new AppError("A team with that name already exists");
+    }
     
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
